microservice/requirements.txt
flask == 1.1.2
pandas == 1.2.4
python_dotenv == 0.19.0
scikit-learn == 1.0.2
